requirements.txt
google-images-search
python-dotenv
